added filters and acl checks to program and schedule managers corrected error where newly uploaded schedules were not being set to active added function commentary and corrected style errors in tables files changed fields to required for programs because joomla uses table null handling for asset null handling

+++ com_thm_organizer/admin/views/program_manager/view.html.php
@@ -48,17 +48,23 @@ class THM_OrganizerViewProgram_Manager extends THM_CoreViewList
     protected function addToolBar()
     {
         JToolbarHelper::title(JText::_('COM_THM_ORGANIZER_PROGRAM_MANAGER_VIEW_TITLE'), 'organizer_degree_programs');
-        JToolbarHelper::addNew('program.add');
-        JToolbarHelper::editList('program.edit');
-        JToolbarHelper::custom(
-                               'program.importLSFData',
-                               'import',
-                               '',
-                               'COM_THM_ORGANIZER_ACTION_IMPORT',
-                               true
-                              );
-        JToolbarHelper::deleteList('COM_THM_ORGANIZER_ACTION_DELETE_CONFIRM', 'program.delete');
-        JToolbarHelper::divider();
-        JToolbarHelper::preferences('com_thm_organizer', '500');
+        if (count(THM_OrganizerHelperComponent::getAccessibleDepartments()))
+        {
+            JToolbarHelper::addNew('program.add');
+            JToolbarHelper::editList('program.edit');
+            JToolbarHelper::custom(
+                'program.importLSFData',
+                'import',
+                '',
+                'COM_THM_ORGANIZER_ACTION_IMPORT',
+                true
+            );
+            JToolbarHelper::deleteList('COM_THM_ORGANIZER_ACTION_DELETE_CONFIRM', 'program.delete');
+        }
+        if ($this->getModel()->actions->{'core.admin'})
+        {
+            JToolbarHelper::divider();
+            JToolbarHelper::preferences('com_thm_organizer', '500');
+        }
     }
 }
